# lib/api-zod/src/generated/types/hoursUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HoursUpdate {
  closed: boolean;
  /** @nullable */
  openHour?: number | null;
  /** @nullable */
  closeHour?: number | null;
}