# lib/api-zod/src/generated/types/menuCategoryBase.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MenuCategoryBase {
  id: number;
  slug: string;
  label: string;
  tagline: string;
  sortOrder: number;
}