# lib/api-zod/src/generated/types/menuCategoryUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MenuCategoryUpdate {
  /** @minLength 1 */
  slug?: string;
  /** @minLength 1 */
  label?: string;
  tagline?: string;
  sortOrder?: number;
}