# lib/api-zod/src/generated/types/menuItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MenuItem {
  id: number;
  categoryId: number;
  name: string;
  price: string;
  description: string;
  /** @nullable */
  image?: string | null;
  sortOrder: number;
}