# lib/api-zod/src/generated/types/sitePhotoUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SitePhotoUpdate {
  /** @minLength 1 */
  url: string;
  alt?: string;
}